requirements.txt
hydra-core==1.3.2
numpy==1.26.3
openai==1.65.2
tqdm==4.66.3
typing_extensions==4.12.2
python-dotenv==1.0.1
tsplib95==0.7.1
scipy==1.15.2
scikit-learn==1.6.1
numba==0.61.2
PyYAML==6.0
jsonschema==4.21.0
# OpenCLIP and dependencies (required to run local open_clip models)
# NOTE: For best results install 'torch' using the official installer that matches your CUDA or CPU setup:
# https://pytorch.org/get-started/locally/
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.0.0
timm>=0.9.0
ftfy>=6.1.1
huggingface-hub>=0.18.0
Pillow>=9.4.0
opencv-python==4.10.0.84